+++ R/css_loader.R
@@ -1,0 +1,9 @@
+
+cssLoader <- function(){
+
+  div(
+    class="lds-circle",
+    div(),
+    spsDepend("css_loading")
+  )
+}