+++ inst/assets/js/sps_cssloader.js
@@ -1,0 +1,42 @@
+
+
+function chooseLoader(type){
+  switch(expression) {
+    case "circle":
+      '<div class="lds-circle"><div></div></div>'
+      break;
+    case "dual-ring":
+      '<div class="lds-dual-ring"></div>'
+      break;
+    case "facebook":
+      '<div class="lds-facebook"><div></div><div></div><div></div></div>'
+      break;
+    case "heart":
+      '<div class="lds-heart"><div></div></div>'
+      break;
+    case "ring":
+      '<div class="lds-ring"><div></div><div></div><div></div><div></div></div>'
+      break;
+    case "roller":
+      '<div class="lds-roller"><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div></div>'
+      break;
+    case "default":
+      '<div class="lds-default"><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div></div>'
+      break;
+    case "ellipsis":
+      '<div class="lds-ellipsis"><div></div><div></div><div></div><div></div></div>'
+      break;
+    case "grid":
+      '<div class="lds-grid"><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div></div>'
+      break;
+    case "hourglass":
+      '<div class="lds-hourglass"></div>'
+      break;
+    case "ripple":
+      '<div class="lds-ripple"><div></div><div></div></div>'
+      break;
+    case "spinner":
+      '<div class="lds-spinner"><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div><div></div></div>'
+      break;
+  }
+}